[BO] stats > product details > (select a product). Graph wrong
stats > product details > (select a product). Graph shows "popular" which is really sales, and "sales" which is actually visits x 100, labeled wrong.

As you can see higher in code there are only 2 layers shown, so not even sure why popular was a thought... since it doesn't have a query even.